--- message.json
@@ -1,1 +1,0 @@
-{"name":"bnxcvbn","password":"xcvbnxvbn"}
