+++ src/http_server/services/battlefield.js
@@ -1,0 +1,3 @@
+export const handleAttackMessage = (data) => {
+  
+}
