+++ src/http_server/services/users.js
@@ -1,12 +1,10 @@
+import { getRoomsLength, updateRoomMessage } from '../services/rooms.js'
+
 let users = []
 let activeUsers = []
 
 const isUserOnline = (name) => {
   return (activeUsers.filter(it => it.name === name).length === 1)
-}
-
-export const getAllUsers = () => {
-  return users
 }
 
 export const deleteActiveUser = async (wsId) => {
@@ -45,18 +43,26 @@ export const broadcast = async (message) => {
   }
 }
 
+export const updateWinnersMessage = () => {
+  const winnersList = users.map(it => {
+    return { name: it.name, wins: it.wins }
+  })
+  const winnersMessage = JSON.stringify({
+      type: "update_winners",
+      data:
+        JSON.stringify(winnersList),
+      id: 0,
+  })
+  return winnersMessage
+}
+
 export const registerUser = async (ws, user) => {
-  // console.log(`WS: ${JSON.stringify(ws)}`)
   const { name, password } = user
   const index = users.length
-  console.log(`registerUser userindex: ${index}, name: ${name}, password: ${password}`)
-  // console.log(`registerUser activeUsers: ${JSON.stringify(activeUsers)}`)
-
   try {
     if (typeof name === 'string' && typeof password === 'string') {
       //check if user already registered
       const filteredUser = users.filter(it => it.name === name)[0]
-      console.log(`filteredUser: ${JSON.stringify(filteredUser)}`)
       //check if password for registered user is correct
       if (filteredUser && filteredUser.password === password) {
         //check if user already online
@@ -75,11 +81,34 @@ export const registerUser = async (ws, user) => {
         const newUser = { index, name, password, wins: 0 }
         users.push(newUser)
         activeUsers.push({ index, name, ws })
-        console.log(`users pushed, activeUsers: ${JSON.stringify(activeUsers.map(it => it.name))}`)
+        await sendRegMessage(ws, user, false, '')
+        await broadcast(updateWinnersMessage())
+        console.log(`getRoomsLength: ${getRoomsLength()}`)
+        if (getRoomsLength() > 0) broadcast(updateRoomMessage())
         return Promise.resolve(newUser)
       }
     } else Promise.reject('login and password should be strings')
   } catch {
     console.log(`error registering user: ${error}`)
+  }
+}
+
+export const sendRegMessage = async (ws, user, error, errorText) => {
+  const { name, index } = user
+  const message = JSON.stringify({
+    type: "reg",
+    data:
+      JSON.stringify({
+        name,
+        index,
+        error,
+        errorText
+      }),
+    id: 0,
+  })
+  try {
+    ws.send(message)
+  } catch (error) {
+    console.log(`Error sending Reg Message: ${error}`)
   }
 }
